Fixed an issue where going to the next page would cause an infinite loop in setting PageNumber

+++ source_code/TypeIT/TypeIT/Utilities/XmlHandler.cs
@@ -109,7 +109,7 @@ namespace TypeIT.Utilities
 
             int avg = int.Parse(XmlHandler.GetElementsFromTags(filePath, tag).FirstOrDefault() ?? "0");
 
-            avg = (int)((avg * (pageCount - 1)) + int.Parse(value)) / (pageCount);
+            avg = (int)(avg * (pageCount - 1) + Double.Parse(value) / pageCount);
 
             XElement statistics = doc.Root?.Elements("Statistics").FirstOrDefault();
